# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "G&L BigCommerce API",
        "description": "G&L Integration API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/createProducts": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "create Products IN Workdrive",
                "description": "create Products IN Workdrive",
                "operationId": "d7e6bb8f35417cdc49eafbd66ca42238",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/checkProducts/{id}": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Check Products in Workdrive",
                "description": "Check if a product exists in Workdrive by ID.",
                "operationId": "a560cf4aa8646f87209a1c639964a19a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the product to check",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/uploadFileCsv": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Upload CSV File to Workdrive",
                "description": "Upload CSV File to Workdrive",
                "operationId": "a26a1f715fe52a8ba2c4884b9bf1c857",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/downloadFileExcel": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Download Excel File from Workdrive",
                "description": "Download Excel File from Workdrive",
                "operationId": "f80ee47413abf94d30cba355f7afb5e0",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "archivo descargado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/downloadFileCsv": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Download Csv File from Workdrive",
                "description": "Download Csv File from Workdrive",
                "operationId": "27047e5ad55fc19bbc372f337aa998d7",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "archivo descargado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/process-excel": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Test API",
                "description": "Process Excel file and convert to CSV format",
                "operationId": "cbf5aa0a947be233d64c22d5de935515",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "API test route works!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/process-bigcommerce": {
            "get": {
                "tags": [
                    "Bigcommerce"
                ],
                "summary": "Process Bigcommerce",
                "description": "Inicia el proceso de Bigcommerce y retorna un mensaje de éxito.",
                "operationId": "processBigcommerce",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Archivo procesado y subido correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Bigcommerce",
            "description": "Bigcommerce"
        }
    ]
}